move scripts to external; add default n_runs and model to check_scripts
