+++ scripts/a2c/a2c.py
@@ -1,9 +1,9 @@
+#! /usr/bin/python3
 import datetime
 import json
 import os
 import random
 import shutil
-from importlib import resources
 from pathlib import Path
 
 import hydra
@@ -14,17 +14,12 @@ import tqdm
 import yaml
 from omegaconf import OmegaConf
 from tensordict import TensorDict
-from tensordict.utils import isin
 from torch.distributions.kl import kl_divergence
-from torchrl.data import (
-    LazyTensorStorage,
-    PrioritizedSampler,
-    SamplerWithoutReplacement,
-    TensorDictMaxValueWriter,
-    TensorDictReplayBuffer,
-)
-from torchrl.envs import ExplorationType, InitTracker, TensorDictPrimer, TransformedEnv
-from torchrl.objectives import ClipPPOLoss
+from torchrl.data import LazyTensorStorage, TensorDictReplayBuffer
+from torchrl.data.replay_buffers.samplers import SamplerWithoutReplacement
+
+from torchrl.envs import InitTracker, TensorDictPrimer, TransformedEnv
+from torchrl.objectives import A2CLoss
 from torchrl.objectives.value.advantages import GAE
 from torchrl.record.loggers import get_logger
 
@@ -32,7 +27,6 @@ from acegen.models import adapt_state_dict, models
 from acegen.rl_env import generate_complete_smiles, SMILESEnv
 from acegen.scoring_functions import custom_scoring_functions, Task
 from acegen.vocabulary import SMILESVocabulary
-
 
 try:
     import molscore
@@ -46,7 +40,7 @@ except ImportError as err:
 
 
 @hydra.main(
-    config_path=str(resources.files("acegen.scripts.ppo")),
+    config_path=".",
     config_name="config_denovo",
     version_base="1.2",
 )
@@ -86,7 +80,7 @@ def main(cfg: "DictConfig"):
                 include=cfg.molscore_include,
             )
             for task in MSB:
-                run_ppo(cfg, task)
+                run_a2c(cfg, task)
         else:
             # Save molscore output. Also redirect output to save_dir
             cfg.molscore = shutil.copy(cfg.molscore, save_dir)
@@ -98,15 +92,15 @@ def main(cfg: "DictConfig"):
                 budget=cfg.total_smiles,
                 output_dir=os.path.abspath(save_dir),
             )
-            run_ppo(cfg, task)
+            run_a2c(cfg, task)
     elif cfg.get("custom_task", None):
         task = Task(custom_scoring_functions[cfg.custom_task], budget=cfg.total_smiles)
-        run_ppo(cfg, task)
+        run_a2c(cfg, task)
     else:
         raise ValueError("No scoring function specified.")
 
 
-def run_ppo(cfg, task):
+def run_a2c(cfg, task):
 
     # Get available device
     device = (
@@ -210,43 +204,28 @@ def run_ppo(cfg, task):
         gamma=cfg.gamma,
         lmbda=cfg.lmbda,
         value_network=critic_training,
-        average_gae=False,
+        average_gae=True,
         shifted=True,
     )
-    loss_module = ClipPPOLoss(
-        actor=actor_training,
-        critic=critic_training,
+    loss_module = A2CLoss(
+        actor_network=actor_training,
+        critic_network=critic_training,
         critic_coef=cfg.critic_coef,
         entropy_coef=cfg.entropy_coef,
-        clip_epsilon=cfg.ppo_clip,
         loss_critic_type="l2",
-        normalize_advantage=True,
-        reduction="none",
     )
 
     # Create data storage
     ####################################################################################################################
 
     buffer = TensorDictReplayBuffer(
-        storage=LazyTensorStorage(cfg.num_envs + cfg.replay_batch_size, device=device),
+        storage=LazyTensorStorage(cfg.num_envs, device=device),
         sampler=SamplerWithoutReplacement(),
         batch_size=cfg.mini_batch_size,
         prefetch=4,
     )
 
-    # Create replay buffer
-    ####################################################################################################################
-
-    storage = LazyTensorStorage(cfg.replay_buffer_size, device=device)
-    experience_replay_buffer = TensorDictReplayBuffer(
-        storage=storage,
-        sampler=PrioritizedSampler(storage.max_size, alpha=0.9, beta=1.0),
-        batch_size=cfg.replay_batch_size,
-        writer=TensorDictMaxValueWriter(rank_key="priority"),
-        priority_key="priority",
-    )
-
-    # Create Optimizer
+    # Create optimizer
     ####################################################################################################################
 
     optim = torch.optim.Adam(
@@ -263,7 +242,7 @@ def run_ppo(cfg, task):
     if cfg.logger_backend:
         logger = get_logger(
             cfg.logger_backend,
-            logger_name="ppo",
+            logger_name="a2c",
             experiment_name=f"{cfg.agent_name}_{task.configs.get('task')}",
             wandb_kwargs={
                 "config": dict(cfg),
@@ -277,12 +256,12 @@ def run_ppo(cfg, task):
     ####################################################################################################################
 
     total_done = 0
+    num_updates = 0
     kl_coef = cfg.kl_coef
-    ppo_epochs = cfg.ppo_epochs
     max_grad_norm = cfg.max_grad_norm
     pbar = tqdm.tqdm(total=cfg.total_smiles)
-    num_mini_batches = (cfg.num_envs + cfg.replay_batch_size) // cfg.mini_batch_size
-    losses = TensorDict({}, batch_size=[cfg.ppo_epochs, num_mini_batches])
+    num_mini_batches = cfg.num_envs // cfg.mini_batch_size
+    losses = TensorDict({}, batch_size=[num_mini_batches])
 
     while not task.finished:
 
@@ -307,7 +286,7 @@ def run_ppo(cfg, task):
         done = data_next.get("done").squeeze(-1)
         pbar.update(done.sum().item())
 
-        # Register smiles lengths and real rewards
+        # Register smiles lengths and real rewards and total generated smiles
         log_info = {}
         total_done += cfg.num_envs
         episode_rewards = data_next["reward"][done]
@@ -323,91 +302,53 @@ def run_ppo(cfg, task):
                 }
             )
 
-        # Get data to be potentially added to the replay buffer later
-        replay_data = data.clone()
+        # Compute advantage
+        with torch.no_grad():
+            data = adv_module(data)
 
-        for j in range(ppo_epochs):
+        # Add extended_data to buffer
+        buffer.extend(data)
 
-            # Compute experience replay loss
-            if (
-                cfg.experience_replay
-                and len(experience_replay_buffer) > cfg.replay_batch_size
-            ):
-                replay_batch = experience_replay_buffer.sample()
-                replay_batch = replay_batch.exclude(
-                    "_weight", "index", "priority", inplace=True
-                )
-                extended_data = torch.cat([data, replay_batch], dim=0)
-            else:
-                extended_data = data
+        for j, batch in enumerate(buffer):
 
-            # Compute advantage
+            batch = batch.to(device, non_blocking=True)
+
+            # Compute loss
+            mask = batch.get("mask").squeeze(-1)
+            loss = loss_module(batch)
+            loss = loss.named_apply(
+                lambda name, value: (
+                    (value * mask).mean() if name.startswith("loss_") else value
+                ),
+                batch_size=[],
+            )
+            loss_sum = (
+                loss["loss_critic"] + loss["loss_objective"] + loss["loss_entropy"]
+            )
+            losses[j] = loss.select(
+                "loss_critic", "loss_entropy", "loss_objective"
+            ).detach()
+
+            # Add KL loss term
             with torch.no_grad():
-                extended_data = adv_module(extended_data)
-
-            # Add extended_data to PPO buffer
-            buffer.extend(extended_data)
-
-            for i, batch in enumerate(buffer):
-
-                # PPO loss
-                mask = batch.get("mask").squeeze(-1)
-                loss = loss_module(batch)
-                loss = loss.named_apply(
-                    lambda name, value: (
-                        (value * mask).mean() if name.startswith("loss_") else value
-                    ),
-                    batch_size=[],
-                )
-                loss_sum = (
-                    loss["loss_critic"] + loss["loss_objective"] + loss["loss_entropy"]
-                )
-                losses[j, i] = loss.select(
-                    "loss_critic", "loss_entropy", "loss_objective"
-                ).detach()
-
-                # Add KL loss term
-                with torch.no_grad():
-                    prior_dist = prior.get_dist(batch)
+                prior_dist = prior.get_dist(batch)
                 kl_div = kl_divergence(actor_training.get_dist(batch), prior_dist)
-                kl_div = (kl_div * mask.squeeze()).sum(-1).mean(-1)
-                loss_sum += kl_div * kl_coef
-                losses[j, i] = TensorDict(
-                    {"kl_div": kl_div.detach().item()}, batch_size=[]
-                )
+            kl_div = (kl_div * mask.squeeze()).sum(-1).mean(-1)
+            loss_sum += kl_div * kl_coef
+            losses[j] = TensorDict({"kl_div": kl_div.detach().item()}, batch_size=[])
 
-                # Update policy
-                loss_sum.backward()
-                torch.nn.utils.clip_grad_norm_(
-                    loss_module.parameters(), max_norm=max_grad_norm
-                )
-                optim.step()
-                optim.zero_grad()
+            # Update policy
+            loss_sum.backward()
+            torch.nn.utils.clip_grad_norm_(
+                loss_module.parameters(), max_norm=max_grad_norm
+            )
+            optim.step()
+            optim.zero_grad()
+            num_updates += 1
 
         losses_mean = losses.apply(lambda x: x.float().mean(), batch_size=[])
         for key, value in losses_mean.items():
             log_info.update({f"train/{key}": value.item()})
-
-        # Add new experiences to the replay buffer
-        if cfg.experience_replay is True:
-
-            # MaxValueWriter is not compatible with storages of more than one dimension.
-            replay_data.batch_size = [replay_data.batch_size[0]]
-
-            # Remove SMILES that are already in the replay buffer
-            if len(experience_replay_buffer) > 0:
-                is_duplicated = isin(
-                    input=replay_data,
-                    key="action",
-                    reference=experience_replay_buffer[:],
-                )
-                replay_data = replay_data[~is_duplicated]
-
-            # Add data to the replay buffer
-            if len(replay_data) > 0:
-                reward = replay_data.get(("next", "reward"))
-                replay_data.set("priority", reward)
-                experience_replay_buffer.extend(replay_data)
 
         if logger:
             for key, value in log_info.items():
