+++ scripts/pretrain/pretrain.sh
