+++ scripts/pretrain/pretrain_single_node.py
@@ -1,3 +1,4 @@
+#! /usr/bin/python3
 import logging
 import os
 import random
