+++ tests/check_scripts/run_pretrain_single_node.sh
@@ -10,6 +10,14 @@
 current_commit=$(git rev-parse --short HEAD)
 project_name="acegen-scripts-check-$current_commit"
 agent_name="pretrain"
+if [ -z "$N_RUN" ]; then
+  echo "N_RUN is not set. Setting to default value of 1."
+  N_RUN=1
+fi
+if [ -z "$ACEGEN_MODEL" ]; then
+  echo "ACEGEN_MODEL is not set. Setting to default value of gru. Choose from [gru, lstm, gpt2]"
+  ACEGEN_MODEL="gru"
+fi
 
 export PYTHONPATH=$(dirname $(dirname $PWD))
 python $PYTHONPATH/scripts/pretrain/pretrain_single_node.py \
@@ -20,7 +28,8 @@ python $PYTHONPATH/scripts/pretrain/pretrain_single_node.py \
   model_log_dir="$agent_name"_seed"$N_RUN" \
   dataset_log_dir="$agent_name"_seed"$N_RUN" \
   epochs=10 \
-  train_dataset_path=$PYTHONPATH/tests/data/smiles_test_set
+  train_dataset_path=$PYTHONPATH/tests/data/smiles_test_set \
+  model=$ACEGEN_MODEL
 
 # Capture the exit status of the Python command
 exit_status=$?
